CLI: remove `--no-logging` flag
The `listen` command at one point supported a mode where it would skip
logging requests in Play's UI.

Today, this doesn't seem to be supported (Play rejects the request
because it says the token is invalid).

Rather than work to repair this in Play, we're just going to remove the
flag.