Libs: change release workflows to be triggered by release published
We had an issue where pushing tags named for version numbers caused
unwanted side-effects.

Attempting to avoid this:
- no longer manually push tags.
- let the GitHub release automatically tag when it publishes.
- drive the other release workflows based on release: published events.

By moving the trigger off of tag push, this additionally opens the door
for situations where we need to push a tag then create a release for it
since the lib workflows won't be triggered until the release is
finalized.