CLI: update codegen for optional request bodies
For the API calls that require a body but all the fields are technically
optional, the ideal case is to allow the CLI command to be invoked
without a body argument at all.

This diff wraps the `JsonOf` fields in this situation with `Option` to
this effect.

The generator template changes were "basic" in so much as the
`Option`-wrapping happens uniformally for cases where passing nothing
isn't especially useful (e.g. `PATCH` request).

With some additional special-casing, we can probably make those required
arguments, but for now they are optional.