CLI: remove top-level Cargo workspace, bump deps, raise dist version
This is some general maintenance around the CLI.

Removing the top-level Cargo.toml helps with rust-analyzer
configuration. Initially it wasn't clear if dist required this
(bootstrapping the dist workspace didn't seem to work without it) but
now that everything is in place, I think we don't need it.

Since there's no longer a top-level `Cargo.lock` I took this opportunity
to `cargo update` the CLI's package-local lockfile.