python: Add openapi-codegen generated integration.py
Breaking change: in the create and update function the integ_(in/update) is renamed to integration_(in/update)

This is due to the way that the codegen decides how to name function arg. (The type of the function arg to_snake_case)
I don't think it makes sense to add a special case in the codegen to shorten `integration` to `integ`
This problem is a bit more annoying, since the path param is called `integ_id` (in the `openapi.json`). So to keep consistency we may want to add the special case to the codegen

Alternatively (and this would make me happy) we could change the path param name from `integ_id` to `integration_id`. (in the `openapi.json`) This will make the file consistent, and won't require a special case in the openapi-codegen