CLI: enable homebrew installer
The dist docs for for the `homebrew` installer indicate there are two
parts to this:

1. create a "tap" repo and a github token for dist to use to manage it
2. configure dist here (with the token as a secret)

The first piece will need to be done by someone else.
This diff does the 2nd piece.

The config for dist targets the extant `svix/homebrew-svix` repo, which
appears to be where the tap lived previously.
It's not clear if we'd need to empty that repo, reboot it, or anything
else.

Refs:
- https://opensource.axo.dev/cargo-dist/book/installers/homebrew.html
- https://github.com/svix/homebrew-svix