Libs: remove beta "events public" and replace with "message poller"
There are new symbol names for the polling endpoint APIs, now no longer
gated by the `svix_beta` cargo feature in the Rust SDK. These are now
available in all Libs.

Relevant updates were made to svix-cli and svix-bridge to account for
the SDK changes, and to ensure user-facing terms are consistent.