docker: convert server dockerfile to use v1 syntax (#2025)
This converts the `server/` Dockerfile to use the newer Dockerfile v1
features (specifically, requiring at least Docker 1.2, introduced in
2020), with things like heredocs and cache mounts. This is the same as
svix/svix-webhooks#2023, but for `server`.

In addition to the aforementioned changes, this also makes `CMD` a
separate bash script instead of a multiline heredoc; this doesn't make
that much of a difference in OSS (`ps` is less ugly, but whatever) but
will let us greatly clean up the private version.