Remove stub for /event-type/schema/generate-example
This is a route that exists in Svix cloud and is being removed. It's
unclear to me why this stub was added here, as it was never exposed in
our SDKs or otherwise used in client software that might work with the
OSS backend rather than cloud, as far as I can tell.