Add no-proxy config support
Allow passing a list of IP address, CIDR ranges, or hostnames
that will bypass the defined proxy server. Hostnames support
basic wildcard syntax, so any hostname ending with `.example.com`
will match the wildcard `*.example.com`.

Note that we do not try to resolve hostnames here, so in effect, IP
address and CIDR ranges will only be relevant for endpoints that
actually use an IP address for the hostname.