Add no-proxy config support (#2146)
Allow passing a list of IP addresses, CIDR ranges, or hostnames that
will bypass the defined
proxy server. Hostnames can be complete or partial (i.e., `.svix.com`
and `svix.com` are
equivalent and will match both `svix.com` and all its subdomains).

Note that we do not try to resolve hostnames prior to determining
whether to route to
the proxy. So in effect, IP addresses and CIDR ranges are only relevant
for endpoints that
actually use an IP address for the hostname.

Fixes #2119